safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/reflected-xss-and-xss-protection-unset-allow-expected.txt

ALERT: XSS
ALERT: Loaded http://127.0.0.1:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cscript%3Ealert(String.fromCharCode(0x58,0x53,0x53))%3C/script%3E&disable-protection=1 into the IFrame.
Testing behavior when "reflected-xss" is set to unset, and "X-XSS-Protection" is set to allow.
